Fix the case where the lowest window is too small to split.
Currently lusty-explorer unconditionally splits the lowest window on the frame.
But split-window can fail for a multitude of reasons, many of which are
associated with the geometry or flags of the currently-open window.  So on
failure, delete the current window and retry the split (failing hard if that
still fails).  (We must delete the current window in this case, since otherwise
lusty-buffer-explorer will get very confused about which window we are trying to
change the current buffer of.  Always creating a new window also means we can
ignore complexities introduced by things like dedicated windows.)