Fix infrequent but long-standing display bug
Account for the absence of a fringe (the case for me in terminal
Emacs). If there is no fringe in a window, the final column of the
window body (which displays the buffer) is repurposed to display the
line continuation character or the line truncation indicator.
Therefore in this configuration we have one fewer columns than Emacs
tells us we do.

This would only manifest if the completions layout algorithm found a
layout that took up exactly as many columns as were available-
generally rare. But it was distracting when it happened!